[#57] refactor: 컨트롤러 로직 반복되는 부분 헬퍼 함수화 및 options 추출
