+++ server/controllers/payment.controller.ts
@@ -1,7 +1,6 @@
 import { Request, Response, NextFunction } from 'express';
 import paymentServices from '../services/payment.services';
-import { ResultRawType } from '../types/types';
-import { getPayload } from '../utils/getPayload';
+import { extractInsertId, getPayload } from '../utils/helper';
 
 class PaymentController {
   async findPayments(req: Request, res: Response, next: NextFunction) {
@@ -25,9 +24,7 @@ class PaymentController {
       const userId = getPayload(req);
       const { type } = req.body;
       const result = await paymentServices.createPayment({ userId, type });
-      const {
-        raw: { insertId },
-      }: ResultRawType = result!;
+      const insertId = extractInsertId(result);
 
       return res.status(200).json({
         insertId,
@@ -42,7 +39,6 @@ class PaymentController {
   async deleteCategory(req: Request, res: Response, next: NextFunction) {
     try {
       const userId = getPayload(req);
-      // 선택한 카테고리의 고유 id 값을 의미
       const { paymentId } = req.params;
       const result = await paymentServices.deleteUserPaymentByUserId({
         userId,
